Move parameter generators to their own module
